Move merge tools to helpers
