+++ kiosk/src/js/helpers/setting_merge_tool.js
